% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{authenticate}
\alias{authenticate}
\title{Invoke the Authentication Process with Bing}
\usage{
authenticate(save = F, path = ".bingauth")
}
\arguments{
\item{save}{logical, should the authentication information should be saved on disk? Defaults to FALSE.}

\item{path}{path to auth file directory}
}
\value{
data.frame containing credentials, optionally saved as .RData in the current working directory
}
\description{
This function starts the authentication process with
Bing. Note that this function needs user interaction.
}
\examples{
\dontrun{
bing_auth <- authenticate(save = F)
}
}
