% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh_token.R
\name{.refresh_token}
\alias{.refresh_token}
\title{Refresh the Access Token}
\usage{
.refresh_token(credentials, refresh_token)
}
\arguments{
\item{credentials}{list of credentials}

\item{refresh_token}{refresh token from previous authentication}
}
\value{
Access token with corresponding time stamp.
}
\description{
refresh_token returns a new valid access token. The access token deprecates after one hour and has to updated with the refresh token.
Usually you need not to run refresh_token() explicitly since the whole authentication process is managed by \code{\link{authenticate}}.
}
