% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Download Reporting Data}
\usage{
download_data(download_url)
}
\arguments{
\item{download_url}{url created with get_report_url}
}
\value{
data
}
\description{
downloads data from download_url
}
\examples{
\dontrun{
data <- download_data(download_url = download_url)
}

}
