% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_report_id.R
\name{get_report_id}
\alias{get_report_id}
\title{Get Report ID}
\usage{
get_report_id(bing_auth, account_id, report, columns, start, end)
}
\arguments{
\item{bing_auth}{auth object generated with authenticate()}

\item{account_id}{account id}

\item{report}{report type}

\item{columns}{columns, attributes and metrics}

\item{start}{start date}

\item{end}{end date}
}
\value{
report id
}
\description{
get_report_id queries the Bing API and request a report id.
}
