% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_bing_data.R
\name{load_bing_data}
\alias{load_bing_data}
\title{Load Bing Data}
\usage{
load_bing_data(bing_auth, account_id, report, columns, start, end)
}
\arguments{
\item{bing_auth}{auth object generated with authenticate()}

\item{account_id}{account id}

\item{report}{report type}

\item{columns}{columns, attributes and metrics}

\item{start}{start date}

\item{end}{end date}
}
\value{
data
}
\description{
load_bing_data queries the Bing Ads API and loads the data.
Note: There is a function timeout after ten seconds if the data download url is not available.
}
\examples{
\dontrun{
data <- load_bing_data(bing_auth = bing_auth,
                       account_id = "*******",
                       report = "CampaignPerformanceReport",
                       columns = c("AccountName", "CampaignName",
                                   "TimePeriod", "Impressions",
                                   "Clicks", "Spend", "Conversions"),
                       start = "2019-01-01",
                       end = "2019-01-31")
}

}
