% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{authenticate}
\alias{authenticate}
\title{Invoke the Authentication Process with Bing}
\usage{
authenticate(save = F, path, gitignore = F)
}
\arguments{
\item{save}{logical, should the authentication information should be saved on disk? Defaults to FALSE.}

\item{path}{path to auth file directory}

\item{gitignore}{logical, adds/appends .gitignore with the authentication information. Defaults to FALSE.}
}
\value{
data.frame containing credentials, optionally saved as .RData in the specified path
}
\description{
This function starts and manages the authentication process with
Bing. If you authenticate for the first time, this function requires user interactions.
You optionally can save the authentication credentials for further usage of the API.
}
\examples{
\donttest{
\dontrun{
if(interactive()){
bing_auth <- authenticate(save = F,
                          path, # ".bingauth"
                          gitignore = F)
}
}
}
}
