% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_download_url.R
\name{get_download_url}
\alias{get_download_url}
\title{Get Report Download URL}
\usage{
get_download_url(bing_auth, report_id)
}
\arguments{
\item{bing_auth}{auth object generated with authenticat()}

\item{report_id}{report id generated by get_report_id()}
}
\description{
get_download_url request the download URL related to the report id generated by get_report_id().
}
\examples{
\dontrun{
download_url <- get_download_url(bing_auth = bing_auth,
                                 report_id = report_id)
}
}
